Refactor interconsultas enviadas y recibidas
