+++ src/pages/api/interconsultas/enviadas/index.ts
@@ -1,9 +1,12 @@
-import { getInterconsultaById } from '@/services/interconsultaSerivce'
+import { filterInterconsultas } from '@/services/interconsultaSerivce'
 import { NextApiRequest, NextApiResponse } from 'next'
 
 interface InterconsultaReq extends NextApiRequest {
   query: {
     id: string
+    estado: string
+    prioridad: string
+    servicio: string
   }
 }
 
@@ -11,8 +14,9 @@ export default async function handler(
   req: InterconsultaReq,
   res: NextApiResponse
 ) {
-  const { id } = req.query
-  const interconsulta = await getInterconsultaById(id)
+  const { id, estado, prioridad, servicio } = req.query
+  const interconsulta = await filterInterconsultas(estado, prioridad, servicio)
+  console.log(id, estado, prioridad, servicio)
   if (!interconsulta) {
     return res
       .status(404)
